backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
scikit-learn==1.5.2
numpy==2.1.1
pydantic==2.8.2
python-dotenv==1.0.0
requests==2.31.0
requests==2.31.0
python-dotenv==1.0.0
